Retire `Bound` in favor of `Index`
This commit introduces an alternative type for indexing sized types like
`vector`. While `Bound` works well enough, it was limited in
expressiveness insofar as it encoded both the bound and indexing value
as a size, thus, one could not create a Vector of bounds that index
different values--it wouldn't type check.

Having the capability to fill a sized functor like Vector with indices
is necessary to define functions such as matrix transposition
generically. So, we've introduced an alternative to Bound, called
`Index` in which only the upper bound is given by a `Size`.

This implementation also simplifies the integer conversions of `Size`,
as we can pass around a `Size` as `Phantom` data.