Add tests for core Vector functions
I've added tests for all of the basic functions over Vectors, `push`,
`tail`, `head`, etc.

In order to do so, I also implemented equality comparison for
Vectors--only Vectors of the same size may be compared.  Two Vectors of
the same size are considered equal if their underlying Arrays of
elements are equal.

This commit also contains a small fix for the `of` macro. Previously,
passing an empty array to the macro resulted in incorrect code. A simple
length check against the argument suffices to prevent this.