Add a doc generation script and Make recipe
The doc generation script will generate documentation for the library in
a docs/ folder. At the moment, there are some hiccups with the
generation (for instance, the logo is not properly respected everywhere)
but I will open a PR against the compiler shortly in order to fix this.

I've also added a docs recipe to generate docs via Make.