Update docs recipe to generate index.html
Since we want to serve the documentation publicly, we need an index page
for visitors to land on. Unfortunately, this lib defines a type called
`Index` and carp uses the module name as the doc's file name.

I've updated the docs recipe to workaround this by renaming the Index
docs and the landing page after generation. We also call on sed to
replace the appropriate links in the generated page index.