Add Hompair type
Hompair is the type of sized pairs of values of a single type, such as
(Int, Int). Just like Vector, it implements the dimensional interfaces
and is transposable with any container type that also implements the
interfaces. Unlike Vector, Hompairs always have a fixed Size of 2.

This commit also updates the documentation accordingly.